refactor: use file-based routing
