+++ src/endpoint/api/apps.ts
@@ -1,0 +1,3 @@
+import type { Context, Env } from 'hono';
+
+export default async function (c: Context<{ Bindings: Env }>) {}
